refactor: rename files, so all have snake case. External headers are untouched
#41
