+++ include/silo/common/input_stream_wrapper.h
