+++ include/silo/query_engine/query_parse_exception.h
@@ -1,5 +1,5 @@
-#ifndef SILO_INCLUDE_SILO_QUERY_ENGINE_QUERYPARSEEXCEPTION_H_
-#define SILO_INCLUDE_SILO_QUERY_ENGINE_QUERYPARSEEXCEPTION_H_
+#ifndef SILO_INCLUDE_SILO_QUERY_ENGINE_QUERY_PARSE_EXCEPTION_H_
+#define SILO_INCLUDE_SILO_QUERY_ENGINE_QUERY_PARSE_EXCEPTION_H_
 
 #include <iostream>
 
@@ -11,4 +11,4 @@ class [[maybe_unused]] QueryParseException : public std::runtime_error {
 };
 }  // namespace silo
 
-#endif  // SILO_INCLUDE_SILO_QUERY_ENGINE_QUERYPARSEEXCEPTION_H_
+#endif  // SILO_INCLUDE_SILO_QUERY_ENGINE_QUERY_PARSE_EXCEPTION_H_
