+++ src/silo/common/input_stream_wrapper.cpp
@@ -1,4 +1,4 @@
-#include "silo/common/InputStreamWrapper.h"
+#include "silo/common/input_stream_wrapper.h"
 
 #include <boost/iostreams/filter/lzma.hpp>
 #include <boost/iostreams/filtering_stream.hpp>
