+++ src/silo/query_engine/query_parse_exception.cpp
@@ -1,4 +1,4 @@
-#include "silo/query_engine/QueryParseException.h"
+#include "silo/query_engine/query_parse_exception.h"
 
 namespace silo {
 [[maybe_unused]] QueryParseException::QueryParseException(const std::string& error_message)
